♻️(jest) Divide code into multiple files
